+++ src/shared/ui/Modal/ModalContext.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+interface ModalContextType {
+  onClose: () => void;
+}
+
+export const ModalContext = React.createContext<ModalContextType | null>(null);
